Change resource name from multi_rel to autodependency
This thought came from the fact that what this is really implementing is a form of
in-language autorequire. But autorequire is really also autobefore, generalized to
autodependency.

It's not referring to the fact that an anchor graph node is used to accomplish the
use case, but it does basically describe the use case.